Model branch

See merge request marius.mauritz/swtii-cal!19

+++ frontend/lib/src/view/components/appointment_component.dart
@@ -1,3 +1,5 @@
+import 'dart:html';
+
 import 'package:angular/angular.dart';
 import 'package:angular_forms/angular_forms.dart';
 import 'package:angular_router/angular_router.dart';
@@ -21,8 +23,10 @@ class AppointmentComponent implements OnActivate {
 
   AppointmentComponent(this._appointmentService, this._location, this._router);
 
-  Appointment appointment = Appointment(1, "Test Termin", "", "", "", "");
+  Appointment appointment;
   final Location _location;
+  bool deleteControl = false;
+  bool submitted = true;
   final Router _router;
   final AppointmentService _appointmentService;
 
@@ -39,16 +43,22 @@ class AppointmentComponent implements OnActivate {
     }
   }
 
+  void onSubmit() => submitted = true;
+
   /// Methode zum speichern, der änderungen die man im Termin vorgenommen hat
   Future<void> save() async {
     await _appointmentService.update(appointment);
-    goBack();
   }
 
   ///Methode zum löschen von Terminen
   Future<void> delete() async {
-    await _appointmentService.delete(appointment.id);
-   goBack();
+    if (deleteControl==true) {
+      await _appointmentService.delete(appointment.id);
+      goBack();
+    } else {
+      deleteControl = true;
+
+    }
   }
 
   ///Methode, die die übergeordnete ansicht anzeigt
